refactor: move aave logic to sit along the reserves entity
