+++ src/services/aave/subgraph/aave-subgraph.client.ts
@@ -1,5 +1,6 @@
 import axios from 'axios';
 import { jsonToGraphQLQuery } from 'json-to-graphql-query';
+
 import { configService as _configService } from '@/services/config/config.service';
 
 export default class AaveSubgraphClient {
